+++ app/controllers/verify/authentications_controller.rb
@@ -2,7 +2,8 @@ module Verify
   class AuthenticationsController < ApplicationController
     CLAIM_TIMEOUT_LENGTH_IN_MINUTES = 90
 
-    before_action :send_unstarted_claiments_to_the_start
+    include PartOfClaimJourney
+
     skip_before_action :verify_authenticity_token, only: [:create]
 
     # Page where a new Verify authentication request is generated and posted, as
